+++ CHANGELOG.md
@@ -12,8 +12,8 @@ Released on 2018-11-30.
 
 #### Added
 
-- Added `BannerViewClosure`, for adding one or more subviews to the `bannerView` instance of `PageViewController`.
 - Added `ForwardedEventsProtocol`.
+- Added support for a footer view below `Segnify`.
   - Added by [Bart Hopster](https://github.com/barthopster).
 
 ## [1.1.0](https://github.com/nedap/Segnify/releases/tag/1.1.0)